requirements.txt
yt-dlp>=2023.12.30
youtube-transcript-api>=0.6.2
openai-whisper>=20231117
tqdm>=4.66.0